backend/requirements.txt
Flask==2.1.2
Werkzeug==2.2.2
Pillow==9.1.1
ultralytics